Add files via upload
Using the API, this tracker is constantly updating and providing accurate information on the state of Covid-19 cases around the world. In addition, the user has the ability to pick a specific country from a drop-down menu and analyze its numbers, which also populate on a bar chart. Lastly, the page is responsive so it will appear correctly across a variety of platforms. Here is the link to the API I used: https://covid19.mathdro.id/api.